+++ visualization/app/codeCharta/state/store/appSettings/appSettings.reducer.ts
@@ -1,0 +1,8 @@
+import { combineReducers } from "redux"
+import { isPresentationMode } from "./isPresentationMode/isPresentationMode.reducer"
+
+const appSettings = combineReducers({
+	isPresentationMode
+})
+
+export default appSettings