+++ visualization/app/codeCharta/state/store/appSettings/isPresentationMode/isPresentationMode.reducer.ts
@@ -1,0 +1,10 @@
+import { PresentationModeAction, PresentationModeActions } from "./isPresentationMode.actions"
+
+export function isPresentationMode(state: boolean = false, action: PresentationModeAction): boolean {
+	switch (action.type) {
+		case PresentationModeActions.SET_PRESENTATION_MODE:
+			return action.payload
+		default:
+			return state
+	}
+}